+++ packages/component/src/Middleware/ActivityGrouping/ui/PartGrouping/private/PartGroupingActivity.tsx
@@ -143,8 +143,8 @@ function PartGroupingActivity(props: PartGroupingActivityProps) {
 
   const classNames = useStyles(styles);
   const getKeyByActivity = useGetKeyByActivity();
-  const [{ collapsibleGroupingDefaultOpen }] = useStyleOptions();
-  const [isGroupOpen, setIsGroupOpen] = useState(collapsibleGroupingDefaultOpen);
+  const [{ partGroupDefaultOpen }] = useStyleOptions();
+  const [isGroupOpen, setIsGroupOpen] = useState(partGroupDefaultOpen);
 
   const messages = useMemo(
     () => activities.map(activity => getOrgSchemaMessage(activity.entities)).filter(message => !!message),